Initial import
Hi,

If you are a git archeologist or a time traveler, this is the first
version of Connections. A remote machine manager developed for the
GNOME desktop.

You might have noticed that in this commit, the application is
called "Remote". That was me facing one of the three hardest problems
in computer science: naming things. Jakub Steiner, who is a much
more creative person, was the one suggesting we name it "Connections".

** How does the World look like today **

- Americans are choosing their Democratic candidate to face Donald
Trump in general election. The dispute is lead by Bernie Sanders and
Joe Biden.
- The Coronavirus (aka COVID-19) is a threatining pandemic of our time.
A few thousand people have died to this date, and milions are in lockdown
in China and various other countries.
- Climate change is a hot topic of our time. Activits continue pressuring
governments to take immediate action. Swedish young activitist Greta
Thunberg has spoken to the UN, US congress, and European Union paraliament
begging for action.
- The United Kingdom has just left the European Union this year.
- China's Chang'E-4 (CE-4) has successfully landed on the eastern floor of
Von Kármán crater within the South Pole–Aitken Basin, becoming the first
spacecraft in history to land on the Moon’s farside.
- It's been about an year since IBM acquired Red Hat.
- Stem cell transplant makes patient’s HIV 'undetectable': In only the second
case of its kind, a stem cell transplant made a London patient’s HIV
“undetectable.” Doctors reported that he was in remission for 18 months
and had stopped taking HIV drugs. Although experts suggest that it’s too
early to say that he was completely cured of HIV, but this marks a step
closer to finding a cure.
- First-ever image of black hole is unveiled: Captured by the Event Horizon
Telescope, the first ever image of a black hole was released on this day.
Located in the Messier 87 galaxy, the black hole is 500 million trillion
km away from Earth and nearly three million times the size of our planet.
- A fire broke out at the 850-year-old Notre-Dame Cathedral in Paris, France.
Officials suggested that the fire may have been ignited by ongoing renovation
work at the gothic landmark.
- Serial bomb blasts rattle Sri Lankan capital killing more than 250 people
and wounded hundreds other. The Islamic State militant outfit claimed
responsibility for the attacks.
- The first Japanese monarch to abdicate in 200 years, Emperor Akihito
stepped down from the Chrysanthemum Throne, marking the end of the Heisei era.
- Trump crossed over the Korean Demilitarized Zone (DMZ) and set foot in North
Korea (DPRK) in order to shake hands with Kim Jong Un, becoming the first sitting
American president to enter the country.
- The strongest known tropical cyclone to impact the Bahamas, Category 5 storm
Dorian made landfall in the northern Bahamas, with a wind speed of 185 mph (298 kph)
and torrential rains. It weakened considerably by the time it reached the U.S.
and made its way to Canada, before dissipating near Greenland a few days later.
At least 50 lives were lost and nearly 76,000 people were left homeless, as per
U.N. estimates.
- NASA astronauts Jessica Meir and Christina Koch created history by performing
an all-woman spacewalk, as they replaced a battery charge/discharge unit on the
International Space Station.
- An eruption rocked New Zealand’s White Island