properties: Add "Address" entry to Properties dialog
By allowing users to add a human readable name for their connection,
we stop exposing the address in the UI.

For that, we should have an entry in the Properties dialog.

Editing an URI is too much trouble since it would require to refresh
the Connection configs. So we would likely recommend users to simply
delete and create a new connection from scratch.