Rename executable back to "gnome-connections"
Back when I initially proposed this app, we were in doubt whether
it would make sense to have such application being part of the core
set of GNOME apps. For this reason I attempted to remove most of
the GNOME branding from it until a decision was reached.

Now we are a GNOME app and we are better off telling the world
about it. :)

Fixes #24