display-view, vnc: Make sure Vnc.Display always has a window
This introduces a dispose_display() method for Connection objects
to have a way to run protocol-specific instructions when a connection
is supposed to persist but hide its display.

This is useful in particular for Vnc.Display, which always requires
a Gtk.Window mapped to it.

This fixes the issue preventing multiple VNC connections to work
in parallel.