display-view: Fix position of cursor for original scale
The position of cursor was offset by certain amount when
the display had original scale. This was caused by
the commit 7f1e29e17988b964f5c0069bcff3b566e50b6865
(display-view: Account for offset of event_box child when relaying events)
which seems to have been fixed on another place in between.
The issue was present in both backends VNC and RDP.