rdp: Allow to resize remote desktops
Change handling of size of remote desktops similarly to vnc connections.
There is 1 combo box which controls it now. It can have 2 or 3 items.
These are "Resize desktop", "Fit window" and "Original size".
The first one has to be supported by the remote server and usually works for
Windows 8 and newer. It has to be added asynchonously once we know that it is
supported which can happen several seconds after connection.