DisplayView: Fix scrolling in VNC connections
gtk-vnc does not process smooth scroll events because VNC does not support them.
Since RDP supports them we need to have smooth scroll events for RDP but
not for VNC. This commit converts smooth scroll events for display widgets
which does not support them to classic scroll events based on a simple
heuristic.